Rename mltoolbox magic to mlworkbench (#455)
* sw

* sw
