+++ setup.py
@@ -37,8 +37,8 @@ setup(
     'google.datalab.utils',
     'google.datalab.utils.commands',
     'google.datalab.contrib',
-    'google.datalab.contrib.mltoolbox',
-    'google.datalab.contrib.mltoolbox.commands',
+    'google.datalab.contrib.mlworkbench',
+    'google.datalab.contrib.mlworkbench.commands',
     'datalab.bigquery',
     'datalab.bigquery.commands',
     'datalab.context',
