+++ tests/main.py
@@ -39,12 +39,12 @@ import kernel.storage_tests
 import kernel.utils_tests
 import ml.dataset_tests
 import mltoolbox_code_free_ml.all_tests
-import mltoolbox_magic.archive_tests
-import mltoolbox_magic.local_predict_tests
-import mltoolbox_magic.shell_process_tests
 import mltoolbox_structured_data.dl_interface_tests
 import mltoolbox_structured_data.sd_e2e_tests
 import mltoolbox_structured_data.traininglib_tests
+import mlworkbench_magic.archive_tests
+import mlworkbench_magic.local_predict_tests
+import mlworkbench_magic.shell_process_tests
 import stackdriver.commands.monitoring_tests
 import stackdriver.monitoring.group_tests
 import stackdriver.monitoring.metric_tests
@@ -104,9 +104,9 @@ _INTEGRATION_TEST_MODULES = [
     mltoolbox_structured_data.dl_interface_tests,
     mltoolbox_structured_data.sd_e2e_tests,  # Not everything runs in Python 3.
     mltoolbox_structured_data.traininglib_tests,
-    mltoolbox_magic.local_predict_tests,
-    mltoolbox_magic.shell_process_tests,
-    mltoolbox_magic.archive_tests,
+    mlworkbench_magic.local_predict_tests,
+    mlworkbench_magic.shell_process_tests,
+    mlworkbench_magic.archive_tests,
 ]
